requirements.txt
fastapi==0.70.0
uvicorn==0.15.0
pydantic==1.8.2
pyjwt==1.4.2
python-multipart==0.0.5
python-jose==3.3.0
passlib~=1.7.4
jose~=1.0.0
psycopg2-binary==2.9.2